wolfssl: patch cond. compilation for Apple Native Cert Validation
WolfSSL assumes every apple deviecs have access to the Security
Framework, and the functions inside of it is always available whether
the user choose to enable WOLFSSL_APPLE_NATIVE_CERT_VALIDATION or not.

However, one of the functions that is called when using native
validation, `SecTrustEvaluateWithError`, is available only in MacOS
10.14 or later. Since we target MacOS 10.12 or later, this function
isn't availble and thus we need to have it not compiled.

This commit introduces a patch to only enable native validation when
`WOLFSSL_APPLE_NATIVE_CERT_VALIDATION` is enabled.