Add wolfssl fix for chacha20 on aarch64 devices
wolfssl is releasing a fix for chacha20 on aarch64 devices. Since it
takes some time for us to upgrade to the latest wolfssl version, we are
adopting this fix via patching.

The patch is created by the diff between the current wolfssl verion we
are on, i.e. 5.7.4-stable and master.