wolfssl: mlkem codepoint backward compatible patch
From 5.7.6 to 5.8.0, WolfSSL has updated the codepoints of
the postquantum curves according to OQS's specification.
https://github.com/open-quantum-safe/oqs-provider/commit/8abfecd7db7620da0fd76d53739579d115e7bc05#diff-ad6b78182015150daf78fa86f78db241b4489a2137c559c95da15cc62c7c1a7bL89

However, the codepoints are incompatible between versions
at or before 5.7.6 and versions at or after 5.8.0. Hence, this
patch is added to make it backward compatible.

The patch is retrieved from https://github.com/wolfSSL/wolfssl/pull/8827

Behaviour:
If version <= 5.7.6 is connected to version >= 5.8.0, old codepoint
is used.
If version >= 5.8.0 is connected to version >= 5.8.0, new codepoint
is used.